Moved sheets from module/applications/actor to module/actor/applications
Updated references, fixed a type error too
