+++ module/actor/applications/actor-simplified-sheet.ts
@@ -3,12 +3,14 @@ import { ActorSheetGURPS } from './actor-sheet.js'
 import DocumentSheetV2 = foundry.applications.api.DocumentSheetV2
 import { parselink } from '../../../lib/parselink.js'
 
-class NPCSheetGURPS extends ActorSheetGURPS {
-  static override DEFAULT_OPTIONS: DocumentSheetV2.DefaultOptions = {
-    classes: ['npc-sheet', 'sheet', 'actor'],
+class ActorSimplifiedSheetGURPS extends ActorSheetGURPS {
+  static override DEFAULT_OPTIONS: DocumentSheetV2.PartialConfiguration<
+    DocumentSheetV2.Configuration<Actor.Implementation>
+  > &
+    object = {
     position: {
-      width: 750,
-      height: 450,
+      width: 820,
+      height: 900,
     },
   }
 
@@ -17,7 +19,7 @@ class NPCSheetGURPS extends ActorSheetGURPS {
   static override PARTS = {
     main: {
       id: 'sheet',
-      template: 'systems/gurps/templates/actor/npc-sheet-ci.hbs',
+      template: 'systems/gurps/templates/simplified.hbs',
       scrollable: [
         '.gurpsactorsheet',
         '#advantages',
@@ -41,12 +43,8 @@ class NPCSheetGURPS extends ActorSheetGURPS {
     const data = await super._prepareContext(options)
     return {
       ...data,
-      // @ts-expect-error: awaiting type implementation
-      currentdodge: this.actor.system.currentdodge,
-      // @ts-expect-error: awaiting type implementation
-      currentmove: this.actor.system.currentmove,
+      dodge: this.actor.getCurrentDodge(),
       defense: this.actor.getTorsoDr(),
-      parryblock: this.actor.getEquippedParry(),
     }
   }
 
@@ -58,10 +56,6 @@ class NPCSheetGURPS extends ActorSheetGURPS {
   ): Promise<void> {
     await super._onRender(_context, _options)
     const html = $(this.element)
-
-    html.find('.npc-sheet').on('click', ev => {
-      this._onfocus(ev)
-    })
 
     html.find('.rollableicon').on('click', event => this._onClickRollableIcon(event))
   }
@@ -77,4 +71,4 @@ class NPCSheetGURPS extends ActorSheetGURPS {
   }
 }
 
-export { NPCSheetGURPS }
+export { ActorSimplifiedSheetGURPS }
