+++ module/actor/applications/inventory-sheet.ts
@@ -1,10 +1,15 @@
 import { ActorSheetGURPS } from './actor-sheet.js'
+import DocumentSheetV2 = foundry.applications.api.DocumentSheetV2
 
-class ActorCombatSheetGURPS extends ActorSheetGURPS {
-  static override DEFAULT_OPTIONS: foundry.applications.api.DocumentSheetV2.DefaultOptions = {
+class InventorySheetGURPS extends ActorSheetGURPS {
+  static override DEFAULT_OPTIONS: DocumentSheetV2.PartialConfiguration<
+    DocumentSheetV2.Configuration<Actor.Implementation>
+  > &
+    object = {
+    classes: ['npc-sheet', 'sheet', 'actor'],
     position: {
-      width: 670,
-      height: 'auto',
+      width: 700,
+      height: 400,
     },
   }
 
@@ -13,7 +18,7 @@ class ActorCombatSheetGURPS extends ActorSheetGURPS {
   static override PARTS = {
     main: {
       id: 'sheet',
-      template: 'systems/gurps/templates/actor/combat-sheet.hbs',
+      template: 'systems/gurps/templates/inventory-sheet.hbs',
       scrollable: [
         '.gurpsactorsheet',
         '#advantages',
@@ -30,4 +35,4 @@ class ActorCombatSheetGURPS extends ActorSheetGURPS {
   }
 }
 
-export { ActorCombatSheetGURPS }
+export { InventorySheetGURPS }
